+++ react/ComparisonList.tsx
@@ -1,0 +1,3 @@
+import ComparisonList from './components/comparisonList/ComparisonList'
+
+export default ComparisonList